testResourcePurgeDraft() is back in the DTT test
